+++ WearableTrophies/WearTrophy.cs
@@ -7,10 +7,17 @@ public class UseItem
   [HarmonyPriority(Priority.High)]
   static void Prefix(Humanoid __instance, ItemDrop.ItemData item)
   {
+    if (!Helper.IsLocalPlayer(__instance)) return;
+    if (!Configuration.Trophies) return;
     if (item.m_shared.m_itemType != ItemDrop.ItemData.ItemType.Trophy) return;
-    if (ZInput.GetButton("AltPlace"))
+    var vanity = Configuration.IsVanity();
+    if (vanity)
+    {
+      QueuedAction.Vanity = true;
+      QueuedAction.ForceSlot = Configuration.GetForceSlot();
       __instance.ToggleEquipped(item);
-    ContainsItem.ForceNoContain = ZInput.GetButton("AltPlace") || (item.m_equipped && item.m_stack <= 1);
+    }
+    ContainsItem.ForceNoContain = vanity || (item.m_equipped && item.m_stack <= 1);
   }
 
   static void Postfix()
@@ -27,28 +34,16 @@ public class ContainsItem
   static bool Postfix(bool result) => result && !ForceNoContain;
 }
 
-[HarmonyPatch(typeof(ItemDrop.ItemData), nameof(ItemDrop.ItemData.GetTooltip), typeof(ItemDrop.ItemData), typeof(int), typeof(bool), typeof(float))]
-public class GetTooltip
-{
-  static string Postfix(string result, ItemDrop.ItemData item)
-  {
-    if (item.m_shared.m_itemType == ItemDrop.ItemData.ItemType.Ammo) return result;
-    if (item.m_shared.m_itemType != ItemDrop.ItemData.ItemType.Trophy && !item.IsEquipable()) return result;
-    if (VanityEquipment.IsEquipped(item))
-      return result + VanityEquipment.GetTooltip(item);
-
-    result = result + "\n\nHold [<color=yellow><b>$KEY_AltPlace</b></color>] for vanity";
-    if (VanityEquipment.IsWeapon(item))
-      result = result + "\nHold [<color=yellow><b>$KEY_Crouch</b></color>] for left hand";
-    if (VanityEquipment.IsWeapon(item))
-      result = result + "\nHold [<color=yellow><b>$KEY_Sit</b></color>] for back item";
-    return result;
-  }
-}
-
 ///<summary>Makes trophies equipable.</summary>
 [HarmonyPatch(typeof(ItemDrop.ItemData), nameof(ItemDrop.ItemData.IsEquipable))]
 public class IsEquipable
 {
-  static bool Postfix(bool result, ItemDrop.ItemData __instance) => result || __instance.m_shared.m_itemType == ItemDrop.ItemData.ItemType.Trophy && ZInput.GetButton("AltPlace");
+  static bool Postfix(bool result, ItemDrop.ItemData __instance) =>
+    result ||
+    (Configuration.ForceAnyItem && (__instance.m_equipped || Configuration.IsForceVanity())) ||
+    (
+       __instance.m_shared.m_itemType == ItemDrop.ItemData.ItemType.Trophy && (
+          __instance.m_equipped || (Configuration.Trophies && Configuration.IsVanity())
+      )
+    );
 }