+++ WearableTrophies/config/ConfigWrapper.cs
@@ -1,0 +1,61 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using BepInEx.Configuration;
+using ServerSync;
+
+namespace Service;
+
+public class ConfigWrapper
+{
+
+  private readonly ConfigFile ConfigFile;
+  private readonly ConfigSync ConfigSync;
+  public ConfigWrapper(string command, ConfigFile configFile, ConfigSync configSync)
+  {
+    ConfigFile = configFile;
+    ConfigSync = configSync;
+
+    new Terminal.ConsoleCommand(command, "[key] [value] - Toggles or sets a config value.", (Terminal.ConsoleEventArgs args) =>
+    {
+      if (args.Length < 2) return;
+      if (!SettingHandlers.TryGetValue(args[1].ToLower(), out var handler)) return;
+      if (args.Length == 2)
+        handler(args.Context, "");
+      else
+        handler(args.Context, string.Join(" ", args.Args.Skip(2)));
+    }, optionsFetcher: () => SettingHandlers.Keys.ToList());
+  }
+  public ConfigEntry<bool> Bind(string group, string name, bool value, string description)
+  {
+    var configEntry = ConfigFile.Bind(group, name, value, new ConfigDescription(description));
+    Register(configEntry);
+    var syncedConfigEntry = ConfigSync.AddConfigEntry(configEntry);
+    syncedConfigEntry.SynchronizedConfig = true;
+    return configEntry;
+  }
+  private static void AddMessage(Terminal context, string message)
+  {
+    context.AddString(message);
+    Player.m_localPlayer?.Message(MessageHud.MessageType.TopLeft, message);
+  }
+  private readonly Dictionary<string, Action<Terminal, string>> SettingHandlers = [];
+  private void Register(ConfigEntry<bool> setting)
+  {
+    var name = setting.Definition.Key;
+    var key = name.ToLower().Replace(' ', '_');
+    SettingHandlers.Add(key, (Terminal terminal, string value) => Toggle(terminal, setting, name, value));
+  }
+  private static string State(bool value) => value ? "enabled" : "disabled";
+  private static readonly HashSet<string> Truthies = ["1", "true", "yes", "on"];
+  private static bool IsTruthy(string value) => Truthies.Contains(value);
+  private static readonly HashSet<string> Falsies = ["0", "false", "no", "off"];
+  private static bool IsFalsy(string value) => Falsies.Contains(value);
+  private static void Toggle(Terminal context, ConfigEntry<bool> setting, string name, string value)
+  {
+    if (value == "") setting.Value = !setting.Value;
+    else if (IsTruthy(value)) setting.Value = true;
+    else if (IsFalsy(value)) setting.Value = false;
+    AddMessage(context, $"{name} {State(setting.Value)}.");
+  }
+}